Catching Charm (+Critical Capture)
Adds the Catching Charm and the Critical Capture mechanic, adapted to gen II.

There's a comment in the Critical Capture that explains things better. It has a chance to give a bonus catch rate.

The Catching Charm (CATCH. CHARM ingame) is fully functional but is not given by any NPCs.